code changes

+++ 0x12-more_singly_linked_lists/1-listint_len.c
@@ -1,4 +1,5 @@
 #include "lists.h"
+#include <stdio.h>
 
 /**
  * listint_len - returns the number of elements in a linked listint_t list.
@@ -8,7 +9,7 @@
  */
 size_t listint_len(const listint_t *h)
 {
-	unsigned int i;
+	size_t i;
 
 	while (h)
 	{
